+++ src/main/groovy/gui/MatrizTableModel.groovy
@@ -50,7 +50,7 @@ class MatrizTableModel extends AbstractTableModel{
 
     @Override
     void setValueAt(Object aValue, int rowIndex, int columnIndex) {
-        dados.setValueAt(Integer.parseInt(aValue), rowIndex, columnIndex)
+        dados.setValueAt(Float.parseFloat(aValue), rowIndex, columnIndex)
     }
 
     SistemaMatriz getDados() {
